Move ISR propagation thread to ZkIsrManager
