+++ core/src/main/scala/kafka/server/ZkIsrManager.scala
@@ -16,19 +16,33 @@
  */
 package kafka.server
 
-import kafka.utils.{Logging, ReplicationUtils}
+import kafka.utils.{Logging, ReplicationUtils, Scheduler}
 import kafka.zk.KafkaZkClient
 import org.apache.kafka.common.TopicPartition
 import org.apache.kafka.common.protocol.Errors
+import org.apache.kafka.common.utils.Time
 
-class ZkIsrManager(zkClient: KafkaZkClient) extends AlterIsrManager with Logging {
+import java.util.concurrent.TimeUnit
+import java.util.concurrent.atomic.AtomicLong
+import scala.collection.mutable
+
+class ZkIsrManager(scheduler: Scheduler, time: Time, zkClient: KafkaZkClient) extends AlterIsrManager with Logging {
+
+  private val isrChangeNotificationConfig = ReplicaManager.DefaultIsrPropagationConfig
+  // Visible for testing
+  private[server] val isrChangeSet: mutable.Set[TopicPartition] = new mutable.HashSet[TopicPartition]()
+  private val lastIsrChangeMs = new AtomicLong(time.milliseconds())
+  private val lastIsrPropagationMs = new AtomicLong(time.milliseconds())
+
   override def start(): Unit = {
-    // No async processing is done, so nothing to do here
+    scheduler.schedule("isr-change-propagation", maybePropagateIsrChanges _,
+      period = isrChangeNotificationConfig.checkIntervalMs, unit = TimeUnit.MILLISECONDS)
   }
 
   override def clearPending(topicPartition: TopicPartition): Unit = {
     // Since we always immediately process ZK updates and never actually enqueue anything, there is nothing to
-    // clear here so this is a no-op
+    // clear here so this is a no-op. Even if there are changes that have not been propagated, the write to ZK
+    // has already happened, so we may as well send the notification to the controller.
   }
 
   override def enqueue(alterIsrItem: AlterIsrItem): Boolean = {
@@ -38,6 +52,11 @@ class ZkIsrManager(zkClient: KafkaZkClient) extends AlterIsrManager with Logging
       alterIsrItem.leaderAndIsr, alterIsrItem.controllerEpoch)
 
     if (updateSucceeded) {
+      // Track which partitions need to be propagated
+      isrChangeSet synchronized {
+        isrChangeSet += alterIsrItem.topicPartition
+        lastIsrChangeMs.set(time.milliseconds())
+      }
       // Return the given LeaderAndIsr with the new Zk version
       alterIsrItem.callback.apply(Right(alterIsrItem.leaderAndIsr.withZkVersion(newVersion)))
     } else {
@@ -47,5 +66,25 @@ class ZkIsrManager(zkClient: KafkaZkClient) extends AlterIsrManager with Logging
     // Return true since we unconditionally accept the AlterIsrItem. The result of the operation is indicated by the
     // callback, not the return value of this method
     true
+  }
+
+  /**
+   * This function periodically runs to see if ISR needs to be propagated. It propagates ISR when:
+   * 1. There is ISR change not propagated yet.
+   * 2. There is no ISR Change in the last five seconds, or it has been more than 60 seconds since the last ISR propagation.
+   * This allows an occasional ISR change to be propagated within a few seconds, and avoids overwhelming controller and
+   * other brokers when large amount of ISR change occurs.
+   */
+  private[server] def maybePropagateIsrChanges(): Unit = {
+    val now = time.milliseconds()
+    isrChangeSet synchronized {
+      if (isrChangeSet.nonEmpty &&
+        (lastIsrChangeMs.get() + isrChangeNotificationConfig.lingerMs < now ||
+          lastIsrPropagationMs.get() + isrChangeNotificationConfig.maxDelayMs < now)) {
+        zkClient.propagateIsrChanges(isrChangeSet)
+        isrChangeSet.clear()
+        lastIsrPropagationMs.set(now)
+      }
+    }
   }
 }
