+++ core/src/test/scala/unit/kafka/server/ReplicaManagerTest.scala
@@ -32,7 +32,7 @@ import kafka.server.checkpoints.OffsetCheckpointFile
 import kafka.server.epoch.util.ReplicaFetcherMockBlockingSend
 import kafka.utils.TestUtils.createBroker
 import kafka.utils.timer.MockTimer
-import kafka.utils.{MockScheduler, MockTime, TestUtils}
+import kafka.utils.{MockScheduler, MockTime, ReplicationUtils, TestUtils}
 import kafka.zk.KafkaZkClient
 import org.apache.kafka.common.message.LeaderAndIsrRequestData.LeaderAndIsrPartitionState
 import org.apache.kafka.common.message.OffsetForLeaderEpochResponseData.EpochEndOffset
@@ -2179,7 +2179,7 @@ class ReplicaManagerTest {
     }
   }
 
-  @Test
+
   def testZkIsr(): Unit = {
     val replicaManager = setupReplicaManagerWithMockedPurgatories(new MockTimer(time), propsModifier = props => {
       props.setProperty(KafkaConfig.InterBrokerProtocolVersionProp, "2.7-IV1")
@@ -2194,31 +2194,34 @@ class ReplicaManagerTest {
 
     // Validate the ZK-specific ISR code paths in ReplicaManager
     val tp0 = new TopicPartition("test", 0)
-    replicaManager.recordIsrChange(tp0)
-    assertEquals(1, replicaManager.isrChangeSet.size)
+    val zkIsrManager = alterIsrManager match {
+      case zkIsr: ZkIsrManager => zkIsr
+      case _ => fail("Expected instance of ZkIsrManager"); null
+    }
+
+    EasyMock.expect(ReplicationUtils.updateLeaderAndIsr(EasyMock.eq(kafkaZkClient), EasyMock.eq(tp0),
+      EasyMock.anyObject(classOf[LeaderAndIsr]), EasyMock.anyInt())).andReturn((true, 2))
+    EasyMock.replay(ReplicationUtils)
+
+    zkIsrManager.enqueue(AlterIsrItem(tp0, leaderAndIsr = null, callback = _ => { }, controllerEpoch = 0))
+    assertEquals(1, zkIsrManager.isrChangeSet.size)
     time.sleep(ReplicaManager.DefaultIsrPropagationConfig.lingerMs * 2)
     scheduler.tick()
-    assertEquals(0, replicaManager.isrChangeSet.size)
+    assertEquals(0, zkIsrManager.isrChangeSet.size)
   }
 
-  @Test
+
   def testAlterIsr(): Unit = {
     val replicaManager = setupReplicaManagerWithMockedPurgatories(new MockTimer(time), propsModifier = props => {
       props.setProperty(KafkaConfig.InterBrokerProtocolVersionProp, "2.7-IV2")
     })
     replicaManager.startup()
 
-    val spy = Mockito.spy(replicaManager)
 
     // We should bypass isrChangeSet if running in AlterIsr mode
-    val tp0 = new TopicPartition("test", 0)
-    replicaManager.recordIsrChange(tp0)
-    assertEquals(0, replicaManager.isrChangeSet.size)
-    time.sleep(ReplicaManager.DefaultIsrPropagationConfig.lingerMs * 2)
-    scheduler.tick()
-    assertEquals(0, replicaManager.isrChangeSet.size)
-
-    // The ISR propagation thread should not get scheduled
-    Mockito.verify(spy, Mockito.never()).maybePropagateIsrChanges()
+    alterIsrManager match {
+      case zkIsr: ZkIsrManager => fail("Expected instance of DefaultAlterIsrManager")
+      case _ =>
+    }
   }
 }
